Update batch_predict.php
Line 51 needs to have: 
->setInputUris($inputUri);
This is because ->setInputUri does not work with the current version. 

LINE 55 should have: 
->setOutputUriPrefix($outputUri);
instead of 
->setInputUri($outputUri);

Both of these are shown on the GCP documentation page, and do not work.
https://cloud.google.com/vision/automl/docs/predict-batch


This fix was WORKING until recently, but now $operationResponse->operationSucceeded() (line 71) returns false, and $error->getMessage() (line 76) returns NULL. Either way, the documentation is broken until someone updates it further for us.